Added extra textview to display currentword. Now grabbing objects from the object manager to get words to match against. Updated match functionality and moved it into postExecute in the cloudvision uploadImage function.
Added setters to the object class, restored the private qualifier to object class members.

Updated the button looks to be more intuitive: bigger take photo button in the middle of the screen, smaller blue buttons for report and skip. Added the skip button; todo: add functionality to onClick for the skip button.